Print metodu test edildi.

+++ Shopping/Entity/Coupon.cs
@@ -6,13 +6,13 @@ namespace Entity
 {
     public class Coupon : BaseEntity
     {
-        public Coupon(int limit,decimal value,DiscountType type)
+        public Coupon(int limit,double value,DiscountType type)
         {
             Limit = limit;
             Value = value;
             DiscountType = type;
         }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
         public int Limit { get; set; }
         public DiscountType DiscountType { get; set; }
     }
